Test refresh firefox dialog
